Integration length

+++ src/board/nap/track_channel.h
@@ -37,7 +37,7 @@ void nap_track_init(u8 channel,
 void nap_track_update(u8 channel,
                       double doppler_freq_hz,
                       double code_phase_rate,
-                      u32 chips_to_correlate,
+                      double chips_to_correlate,
                       bool has_pilot_sync);
 
 void nap_track_read_results(u8 channel,
