+++ src/track/track_api.c
@@ -67,7 +67,7 @@ bool nap_sc_wipeoff(const tracker_t *tracker) {
  * \param[in]     tracker Tracker channel data
  * \param chips_to_correlate  Number of code chips to integrate over.
  */
-void tracker_retune(tracker_t *tracker, u32 chips_to_correlate) {
+void tracker_retune(tracker_t *tracker, double chips_to_correlate) {
   double doppler_freq_hz = tracker->carrier_freq;
   double code_phase_rate = tracker->code_phase_rate;
   bool nap_strip_sc = nap_sc_wipeoff(tracker);
