+++ src/track/track_api.h
@@ -28,7 +28,7 @@ void tracker_correlations_read(u8 nap_channel,
                                u32 *sample_count,
                                double *code_phase,
                                double *carrier_phase);
-void tracker_retune(tracker_t *tracker, u32 chips_to_correlate);
+void tracker_retune(tracker_t *tracker, double chips_to_correlate);
 s32 tracker_tow_update(tracker_t *tracker,
                        s32 current_TOW_ms,
                        u32 int_ms,
